+++ ProcessMap.py
@@ -2,7 +2,6 @@ from Log import Log
 from TransitionMatrix import TransitionMatrix
 from Graph import Graph
 from Renderer import Renderer
-from Updater import Updater
 
 class ProcessMap():
 
@@ -10,9 +9,9 @@ class ProcessMap():
         self.Log = None
         self.Rates = {'activities': 100, 'paths': 0}
         self.Params = {'optimize': True, 'aggregate': False}
-        self.__Observers = {'T': None,
-                            'Graph': None,
-                            'Renderer': None}
+        self._Observers = {'T': None,
+                           'Graph': None,
+                           'Renderer': None}
 
     def set_log(self, FILE_PATH, c=(0,1), *args, **kwargs):
         log = Log(FILE_PATH, cols=c, *args, **kwargs)
@@ -38,7 +37,7 @@ class ProcessMap():
         UPD.update()
         if self.Params['optimize']:
             self.Rates = UPD.Rates
-        self.__Observers = UPD.Observers
+        self._Observers = UPD._Observers
 
     def get_log(self):
         return self.Log
@@ -50,10 +49,46 @@ class ProcessMap():
         return self.Params
 
     def get_T(self):
-        return self.__Observers['T']
+        return self._Observers['T']
 
     def get_graph(self):
-        return self.__Observers['Graph']
+        return self._Observers['Graph']
 
     def render_map(self, save_path='', colored=True):
-        return self.__Observers['Renderer']
+        return self._Observers['Renderer']
+
+
+class Updater(ProcessMap):
+
+    def get_T(self):
+        T = TransitionMatrix()
+        T.update(self.Log.flat_log)
+        return T
+
+    def get_graph(self):
+        G = Graph()
+        if self.Params['optimize']:
+            self.Rates = G.optimize(self.Log,
+                                    self._Observers['T'])
+        else:
+            G.update(self.Log,
+                     self.Rates['activities'],
+                     self.Rates['paths'],
+                     self._Observers['T'])
+        if self.Params['aggregate']:
+            G.aggregate(self.Log,
+                        self.Rates['activities'],
+                        self.Rates['paths'],
+                        self._Observers['T'])
+        return G
+
+    def render_map(self, save_path='', colored=True):
+        R = Renderer()
+        R.update(self._Observers['T'],
+                 self._Observers['Graph'])
+        return R
+
+    def update(self):
+        self._Observers['T'] = self.get_T()
+        self._Observers['Graph'] = self.get_graph()
+        self._Observers['Renderer'] = self.render_map()
